third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/offsetWidth-host-style-change-expected.txt

CONSOLE ERROR: line 12: Uncaught TypeError: host.createShadowRoot is not a function
offsetWidth of a fixed width element should cause a style recalc if host styles are invalid

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE
